WIP finished local test for find_hung_process.py, added files to testers directory
For find_hung_process.py added an alternate qstat command that runs the mock qstat file without adding it to path. It simply swaps out the command that the program runs if the -qp arg is passed in.

Refactored some arg parse commands and added a sub parser for the local testing. To use the "testing mode" of the program, type in "test" in the command line as one of the options.

Added new files FindHungPRocessTester.py and qstat. These will be used for setting up the actual test for the program.

The goal is to eventually be able to run the program locally with the tester and verify that the output is as expected.