WIP Minimum tester set up for find_hung_process.py
They very very bare minimum for the find_hung_process.py tester.

Needed to add a static method, and a "tests" input file for run_tests to work.

Needed to add a simple_diffusion.i file as well for the tester to run the test.

Also added a temp directory called "tester" for the "tests" file. It is in test/tests/tester/tests