Fixed logic and added execute permission for binary
For now the tester will build up strings from an empty string. With how this is used it is okay for now. This can't take in input from the command line currently. Could add this in the future.

PR kept failing civet tests because it could not run the binary. Used chmod to fix this temporarily until better methods are used.